Complete flex exercise 04-flex-information
Initially completed the exercise to match the
desired outcome. However it did not shrink
correctly when reducing the viewport.

I had created an extra container to house the
title and info in a flex column. This step was
unnecessary complication. Removed this extra
container and treated the title as a block item.

This left just 1 container, which housed 4
children info 'cards'. This shrinks correctly.