Complete flex exercise 05-flex-modal
Got close to a matching layout quite quickly.
However, when checking the solution, struggled to
understand the button positioning inside the
header div.
After some trialling, managed to match the layout
again, but I preferred the button as part of the
main container, not as part of the header.
Kept the styles.css changes however, as they more
directly matched the desired outcome.